+++ NEWS.md
@@ -1,5 +1,9 @@
 # version 0.5-4
 
+* `c.stars()` fails if it tries to merge arrays with different units; #457
+
+* For NetCDF files, `read_stars()` uses the `long_name` as array name; #457
+
 * add `rename()` method; #470
 
 * refresh CRS of packaged `L7_ETMs.tif`; #466